# [Enhancement] support cast string to datetime cache