Fixes #issue
1. string cast to date/datetime, add cache

from ssb data
```
CREATE TABLE `s5` (
  `v1` int(11) NOT NULL COMMENT "",
  `v2` string NULL,
  `v3` datetime NULL,
  `v4` string NULL,
  `v5` string NULL,
  `v6` string NULL,
  `v7` string NULL
) ENGINE=OLAP
DUPLICATE KEY(`v1`)
COMMENT "OLAP"
DISTRIBUTED BY HASH(`v1`) BUCKETS 48
PROPERTIES (
"replication_num" = "1",
"in_memory" = "false",
"enable_persistent_index" = "false",
"replicated_storage" = "true",
"compression" = "LZ4"
);

insert into s5
select lo_orderdate, cast(cast(lo_orderdate as datetime) as string), cast(lo_orderdate as datetime), l, date_trunc('quarter', l), date_trunc('month', l), date_trunc('week', l)
from (select lo_orderdate, cast(lo_commitdate*1000000 + lo_quantity + (lo_orderkey % 60) * 100 + (lo_custkey % 24) * 10000  as datetime) as l, 
from lineorder) x

select count(cast(v4 as datetime)) from s5;
select count(cast(v2 as datetime)) from s5;
select count(cast(v7 as datetime)) from s5;
select count(cast(v5 as datetime)) from s5;
select count(cast(v6 as datetime)) from s5;
```

### enable low_cardinality_optimize

| s5(600037902) without date partition | count() | count(cast) before | after | before/after |
| ---- | ---- | ---- | ---- | ---- |
| v3: distinct 2406 datetime baseline | 332ms | 337ms | 332ms | 1.02 |
| v4: distinct 115353638 | 1344ms | 5623ms  |  8143ms  | 0.69 |
| v2: distinct 2406      | 653ms |  5415ms  |  2444ms  | 2.22 |
| v7: distinct 353       | 688ms |  5305ms  |  2762ms  | 1.92 |
| v6: distinct 82        | 275ms |  519ms  |  489ms  | 1.06 |
| v5: distinct 28        | 274ms |  493ms  |  479ms  | 1.03 |

| s5(600037902) with date partition | count() | count(cast) before | after | before/after |
| ---- | ---- | ---- | ---- | ---- |
| v3: distinct 2406 datetime baseline | 329 | 332ms | 325ms | 1.02 |
| v4: distinct 115353638 | 1336ms |  5142ms  |  7427ms  | 0.69 |
| v2: distinct 2406     | 658ms |  4728ms  |  2266ms  | 2.09 |
| v7: distinct 353     | 721ms |  4844ms  |  2583ms  | 1.88 |
| v6: distinct 82       | 291ms |  489ms  |  473ms  | 1.03 |
| v5: distinct 28       | 293ms |  491ms  |  472ms  | 1.04 |


### disable low_cardinality_optimize
```
set cbo_enable_low_cardinality_optimize=false;
 ```

| s5(600037902) without date partition | count() | count(cast) before | after | before/after |
| ---- | ---- | ---- | ---- | ---- |
| v3: distinct 2406 datetime baseline | 381ms | 384ms | 332ms | 1.16 |
| v4: distinct 115353638 | 1344ms | 5668ms  |  8107ms  | 0.70 |
| v2: distinct 2406      | 653ms |  5256ms  |  2670ms  | 1.97 |
| v7: distinct 353       | 688ms |  5223ms  |  3028ms  | 1.72 |
| v6: distinct 82        | 275ms |  5348ms  |  3265ms  | 1.64 |
| v5: distinct 28        | 274ms |  5354ms  |  2834ms  | 1.89 |

| s5(600037902) with date partition | count() | count(cast) before | after | before/after |
| ---- | ---- | ---- | ---- | ---- |
| v3: distinct 2406 datetime baseline | 375ms | 341ms | 380ms | 0.90 |
| v4: distinct 115353638 | 1344ms | 5142ms  |  7550ms  | 0.68 |
| v2: distinct 2406      | 653ms |  4794ms  |  2365ms  | 2.03 |
| v7: distinct 353       | 688ms |  4710ms  |  2699ms  | 1.75 |
| v6: distinct 82        | 275ms |  4795ms  |  3037ms  | 1.58 |
| v5: distinct 28        | 274ms |  4756ms  |  2638ms  | 1.80 |


## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [x] 3.1
  - [x] 3.0
  - [ ] 2.5