@wanpengfei-git 2023-09-15
#### [BE Incremental Coverage Report]

:heart_eyes: **pass** : 25 / 31 (80.65%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|src/exprs/cast_expr.cpp|25|31|80.65%|[851, 852, 853, 958, 959, 960]|